+++ src/main/java/com/skilllease/services/JobBidService.java
@@ -3,7 +3,6 @@ package com.skilllease.services;
 import com.skilllease.dto.JobBidRequestDTO;
 import com.skilllease.entities.JobBid;
 import com.skilllease.exception.AppException;
-import com.skilllease.exception.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 
 import java.util.List;
@@ -13,13 +12,13 @@ public interface JobBidService {
     @Transactional
     JobBid createJobBid(JobBidRequestDTO dto) throws AppException;
 
-    Optional<JobBid> getJobBidById(Integer id);
+    Optional<JobBid> getJobBidById(Long id);
 
     List<JobBid> getJobBidsByJob(Long jobId);
 
     @Transactional
-    JobBid updateJobBidStatus(Integer bidId, String newStatus) throws AppException;
+    JobBid updateJobBidStatus(Long bidId, String newStatus) throws AppException;
 
     @Transactional
-    void deleteJobBid(Integer bidId) throws AppException;
+    void deleteJobBid(Long bidId) throws AppException;
 }
